+++ app/routes/app.cards.tsx
@@ -3,9 +3,9 @@ import { Card, Text, Button, Grid, Spacer, Divider } from "@geist-ui/react";
 import { CreditCard as CreditCardIcon, Eye, EyeOff } from '@geist-ui/react-icons';
 import { useLoaderData } from "@remix-run/react";
 import { LoaderFunction, json } from '@remix-run/cloudflare';
-import { getPrismaClient } from "../util/db.server";
+import { getPrismaClient } from "../service/db.server";
 import { getUserSession } from '../auth.server';
-import { openCard } from "~/util/cardUtil";
+import { openCard } from "~/service/cardService";
 import { CreditCard, DebitCard } from "@prisma/client";
 import ResizableText from '~/components/ResizableText';
 import { useSelector } from 'react-redux';
